Update to the upstream PR #27309
